% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{bind_attrs}
\alias{bind_attrs}
\title{Bind a list of \code{stars} objects by attribute.}
\usage{
bind_attrs(x)
}
\arguments{
\item{x}{list of one or more \code{stars} objects}
}
\value{
\code{stars} objects
}
\description{
Bind a list of \code{stars} objects by attribute.
}
\seealso{
[stars issue 440](https://github.com/r-spatial/stars/issues/440#issuecomment-877627732)
}
