% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{bind_stars}
\alias{bind_stars}
\title{Bind a list of \code{stars} objects.}
\usage{
bind_stars(x, ...)
}
\arguments{
\item{x}{list of one or more \code{stars} objects}

\item{...}{arguments for \code{\link[stars]{c.stars}}}
}
\value{
\code{stars} objects
}
\description{
Bind a list of \code{stars} objects.
}
\seealso{
\code{\link[stars]{c.stars}}
}
