% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{closest_available_cell}
\alias{closest_available_cell}
\title{Determine the closest non-NA pixel given [lon,lat] and a lut raster}
\usage{
closest_available_cell(x = volcano_points(), lut = make_raster_lut())
}
\arguments{
\item{x}{sf POINT data, data.frame or with [lon,lat] coordinates in that order 
and assumed to be aligned with the CRS of the \code{lut}}

\item{lut}{stars object, look-up with precomputed closest non-NA cells. 
See \code{\link{make_raster_lut}}}
}
\value{
sf object with two attributes \code{index} (the index of the closets non-NA point)
  and \code{original} (the original index)
}
\description{
Determine the closest non-NA pixel given [lon,lat] and a lut raster
}
