% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{colrow_from_cells}
\alias{colrow_from_cells}
\title{Given cell indices, convert to col-row addresses}
\usage{
colrow_from_cells(x, cells)
}
\arguments{
\item{x}{stars object}

\item{cells}{num, cell addresses}
}
\value{
matrix of [col, row]
}
\description{
Given cell indices, convert to col-row addresses
}
