% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{combine_ready}
\alias{combine_ready}
\title{A diagnostic tool for comparing the dimensionality of two stars objects}
\usage{
combine_ready(x, y, tolerance = 0, ignore_resolution = FALSE)
}
\arguments{
\item{x}{`stars` or `dimensions` object}

\item{y}{another `stars` or `dimensions` object}

\item{tolerance}{num, see [base::all.equal]}

\item{ignore_resolution}{logival, if TRUE ignore resolution}
}
\value{
logical, TRUE if the objects are compatable with [stars::c.stars]
}
\description{
A diagnostic tool for comparing the dimensionality of two stars objects
}
