% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylength.R
\name{day_length.default}
\alias{day_length.default}
\title{day_length default method}
\usage{
\method{day_length}{default}(x, ...)
}
\arguments{
\item{x}{num latitude values}

\item{...}{Arguments passed to or from other methods}
}
\description{
day_length default method
}
