% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylength.R
\name{day_length}
\alias{day_length}
\title{day_length generic}
\usage{
day_length(x, ...)
}
\arguments{
\item{x}{vector of latitudes, sf or stars object}

\item{...}{Arguments passed to or from other methods}
}
\description{
day_length generic
}
