% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylength.R
\name{day_length.sfc}
\alias{day_length.sfc}
\title{Function to compute daylength for a sfc object}
\usage{
\method{day_length}{sfc}(x = read_buoys(), doy = as.Date("2014-01-01"))
}
\arguments{
\item{x}{sfc object}

\item{doy}{numeric or Date, day-of-year or Date, the date to compute day length}
}
\value{
vector of day length values
}
\description{
Function to compute daylength for a sfc object
}
