% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylength.R
\name{daylength}
\alias{daylength}
\title{Function to create a day length stars object}
\usage{
daylength(
  x = dplyr::slice(stars::read_stars(system.file("datasets/20140601-20140630-sst.tif",
    package = "twinkle")), "band", 1),
  doy = as.Date("2014-01-01"),
  transfer_na = FALSE
)
}
\arguments{
\item{x}{stars, template stars object on which to build}

\item{doy}{numeric or Date, day-of-year or Date, the date to compute day length}

\item{transfer_na}{logical, if TRUE transfer NAs from x to the output}
}
\value{
stars object with day length in hours per cell
}
\description{
Day length is compute form the latitude coordinates for the center of each cell
}
