% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{delta_stars}
\alias{delta_stars}
\title{Retrieve the deltas of stars dimensions}
\usage{
delta_stars(x)
}
\arguments{
\item{x}{stars object}
}
\value{
named list of deltas
}
\description{
Retrieve the deltas of stars dimensions
}
