\name{forex}
\docType{data}
\alias{forex}
\title{data: 8 currency pairs}
\description{
The xts dataset consists of the daily value of 8 currency pairs spanning the 
period \sQuote{31-12-1979} to \sQuote{31-12-1998} from the van Dijk and Franses 
book, and originally sourced from the Federal Reserve Bank of New York.
}
\usage{data(forex)}
\format{An xts matrix containing 4959x8 observations.}
\source{\url{http://people.few.eur.nl/djvandijk/nltsmef/nltsmef.htm}}
\references{
Franses, P. H., and Van Dijk, D. 2000, Non-linear time series models 
in empirical finance, \emph{Cambridge University Press}.
}
\keyword{datasets}
