% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{is_stars}
\alias{is_stars}
\title{Test is an object inherits from \code{stars} class}
\usage{
is_stars(x, class = "stars")
}
\arguments{
\item{x}{object to test}

\item{class}{character, the class to test against, by default 'stars'}
}
\value{
logical
}
\description{
Test is an object inherits from \code{stars} class
}
