% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{mask_stars}
\alias{mask_stars}
\title{Mask a stars object with a polygon}
\usage{
mask_stars(x, y, invert = FALSE, filename = NA)
}
\arguments{
\item{x}{stars object}

\item{y}{sf polygon object.  Pixels exterior of this as assigned NA}

\item{invert}{logical, if TRUE values inside the polygon are set to NA}

\item{filename}{character or NA, if a filename write to the file}
}
\value{
a masked stars object, areas outside (or possibly inside) are set to NA
  depending upon the value of \code{invert}.  Default is outside is NA
}
\description{
You can use st_crop(...) but it runs slowly for large rasters
}
\examples{
\dontrun{
  library(stars)
  library(sf)
  x = read_stars(system.file("tif/L7_ETMs.tif", package = "stars")) 
  m = cbind(c(290000, 295000, 292500, 290000),
            c(9113000, 9115000, 9118000, 9113000))
  p = st_sfc( st_polygon(list(m)) ) |> 
    st_set_crs(31985)
  m = mask_stars(x, p)
  plot_p = function() plot(p, add = TRUE, border = 'orange')
  z = c(x,m, along = 'band')
  plot(z, hook = plot_p)
}
}
\seealso{
https://github.com/r-spatial/stars/issues/622
}
