% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{mplot}
\alias{mplot}
\title{Plot multiple attributes of a stars object}
\usage{
mplot(
  x = volcano_multi(threshold = 120),
  key.pos = NULL,
  mfrow = c(1, length(x)),
  band = 1,
  ...
)
}
\arguments{
\item{x}{stars object with one or more attributes}

\item{key.pos}{see \code{\link[stars]{plot}}}

\item{mfrow}{see \code{\link[graphics]{par}}}

\item{band}{see \code{\link[stars]{plot}}}

\item{...}{other arguments for \code{\link[stars]{plot}}}
}
\description{
This is pretty limited, and legends (keys) are disabled.
}
\examples{
\dontrun{
v = volcano_single(threshold = 120)
lut = make_raster_lut(v)
x = c(v, lut)
mplot(v)
}
}
\seealso{
https://github.com/r-spatial/stars/issues/374#issuecomment-761773542
}
