\name{ndx}
\docType{data}
\alias{ndx}
\title{data: OHLCV for the Nasdaq 100 index}
\description{
The xts dataset consists of the daily open, high, low, close and volume of the 
Nasdaq 100 index spanning the period \sQuote{02-01-1996} to \sQuote{12-10-2001} 
from Yahoo finance.
}
\usage{data(ndx)}
\format{An xts matrix containing 1457x5 observations.}
\source{\url{http://finance.yahoo.com}}
\keyword{datasets}
