% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs.R
\name{pairs,data.frame-method}
\alias{pairs,data.frame-method}
\title{Pairwise matrix of scatterplot for data.frame objects - modeled after [terra::pairs()]}
\usage{
\S4method{pairs}{data.frame}(x, hist = TRUE, cor = TRUE, use = "pairwise.complete.obs", ...)
}
\arguments{
\item{x}{the coordinates of points given as numeric columns of a
    matrix or data frame.  Logical and factor columns are converted to
    numeric in the same way that \code{\link{data.matrix}} does.
  }

\item{...}{
    arguments to be passed to or from methods.

    Also, \link[graphics]{graphical parameters} can be given as can arguments to
    \code{plot} such as \code{main}.  \code{par("oma")} will be set
    appropriately unless specified.
  }
}
\value{
a pairs plot
}
\description{
I first made a `pairs-stars` for the [tidysdm::pairs()] function to match that 
offered by [terra::pairs()].  But we like the plot so much we thought it worth
adding a data.frame method.
}
\examples{
x = iris |>
  dplyr::as_tibble() |>
  dplyr::select(-dplyr::all_of("Species"))
pairs(x)
}
