% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{points_to_convexhull}
\alias{points_to_convexhull}
\title{Compute the convex hull of a set of points}
\usage{
points_to_convexhull(x)
}
\arguments{
\item{x}{sf object with geometry of type POINT}
}
\value{
sf POLYGON table
}
\description{
Compute the convex hull of a set of points
}
\examples{
\dontrun{
pts <- sf::read_sf(system.file("datasets/penbay-points.gpkg", package = 'twinkle'))
ch <- points_to_mesh(pts)
sst <- stars::read_stars(system.file("datasets/20140601-20140630-sst.tif", package = "twinkle"))
plot(sst[,,,1],
     main = "sst 2014-01-01",
     reset = FALSE,
     add.geom = list(sf::st_geometry(pts), col = 'blue', pch = 19))
plot(sf::st_geometry(ch), add = TRUE, border = 'orange', lwd = 2, col = NA)
}
}
\seealso{
\href{https://stackoverflow.com/questions/51718839/in-r-how-do-i-run-st-convex-hull-function-on-point-sf-object}{stackoverflow}
}
