% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{points_to_mesh}
\alias{points_to_mesh}
\title{Convert sfc_POINTS to a mesh of polygons (delaunay triangulation)}
\usage{
points_to_mesh(
  x,
  method = c("geometry", "sf")[1],
  varname = NULL,
  fun = mean,
  ...
)
}
\arguments{
\item{x}{sf object with geometry of type POINT}

\item{method}{charcater, 'geometry' (default) or 'sf'}

\item{varname}{character one or more variable names to transfer to the polygon
Ignored if method is not 'geometry'}

\item{fun}{function to compute value the variables specifed by varname
Ignored if method is not 'geometry'}

\item{...}{other arguments for \code{fun}}
}
\value{
sf POLYGON table is method is 'geometry' and sf GEOMETRYCOLLECTION otherwise
}
\description{
Convert sfc_POINTS to a mesh of polygons (delaunay triangulation)
}
\examples{
\dontrun{
pts <- sf::read_sf(system.file("datasets/penbay-points.gpkg", package = 'twinkle'))
mesh <- points_to_mesh(pts)
sst <- stars::read_stars(system.file("datasets/20140601-20140630-sst.tif", package = "twinkle"))
plot(sst[,,,1],
     main = "sst 2014-01-01",
     reset = FALSE,
     add.geom = list(sf::st_geometry(pts), col = 'orange', pch = 19))
plot(sf::st_geometry(mesh), add = TRUE, border = 'green')
}
}
\seealso{
\code{\link[geometry]{delaunayn}}

\href{https://stackoverflow.com/questions/57677412/sf-how-to-get-back-to-multipolygon-from-geometrycollection}{stackoverflow}
}
