% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoy.R
\name{read_buoys}
\alias{read_buoys}
\title{Read a selection of buoy data}
\usage{
read_buoys()
}
\value{
a sf table of buoy metadata
}
\description{
Data are a subset of the buoys available from NOAA-EDAB group's 
[buoydata](https://github.com/NOAA-EDAB/buoydata) R package.
}
