% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{set_names}
\alias{set_names}
\title{Set names of an object}
\usage{
set_names(x, nm = paste0("v", seq_len(length(x))))
}
\arguments{
\item{x}{object with name-able elements}

\item{nm}{character vector of names, by default \code{v1, v2, v3, ...}}
}
\value{
the input object with names
}
\description{
Set names of an object
}
\seealso{
\code{\link[stats]{setNames}}
}
