% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{sf_column}
\alias{sf_column}
\title{Retrieve the attribute that names to sf column}
\usage{
sf_column(x)
}
\arguments{
\item{x}{sf object}
}
\value{
character, the name of the sf column
}
\description{
Retrieve the attribute that names to sf column
}
