% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{shape_stars}
\alias{shape_stars}
\title{Assemble a vector of stars dimension or shape}
\usage{
shape_stars(x)
}
\arguments{
\item{x}{stars object}
}
\value{
numeric vector of
\itemize{
  \item{ncol the number of columns}
  \item{nrow the number of rows}
  \item{ncell the number of cells (pixels) in each band (ncol * nrow)}
  \item{nband the number of bands}
  \item{nindex the number of indices (ncell*nband)}
  \item{natt the number of attributes}
}
}
\description{
Assemble a vector of stars dimension or shape
}
\examples{
\dontrun{
v <- volcano_multi(what = 'bands')
shape_stars(v)
v2 <- bind_stars(v, v)
shape_stars(v2)
}
}
