% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{st_rotate}
\alias{st_rotate}
\title{Rotate a simple feature}
\usage{
st_rotate(x, r = pi/2, degrees = FALSE)
}
\arguments{
\item{x}{sf object}

\item{r}{numeric, radians (or degrees) to rotate}

\item{degrees}{logical, if TRUE intepret \code{r} as degrees instead of radians}
}
\value{
a rotated version of the input \code{x}
}
\description{
Rotate a simple feature
}
\examples{
\dontrun{
 v <- volcano_multi(what = 'bands')
 x <-  volcano_polygon()
 x2 <- st_rotate(x, -pi/8)
 plot(v[,,,1], reset = FALSE)
 plot(sf::st_geometry(x), add = TRUE, border = "purple", col = NA, lwd = 3)
 plot(sf::st_geometry(x2), add = TRUE, border = "orange", col = NA, lwd = 3)
}
}
\seealso{
https://r-spatial.github.io/sf/articles/sf3.html#affine-transformations-1
}
