% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{st_translate}
\alias{st_translate}
\title{Translate a simple feature}
\usage{
st_translate(x, d = c(0, 0))
}
\arguments{
\item{x}{sf object}

\item{d}{numeric, values to translate by in order of [x, y]}
}
\value{
a translated version of the input \code{x}
}
\description{
Translate a simple feature
}
\examples{
\dontrun{
  v <- volcano_multi(what = 'bands')
  x <-  volcano_polygon()
  x2 <- st_translate(x, c(-250, 120))
  plot(v[,,,1], reset = FALSE, axes = TRUE)
  plot(sf::st_geometry(x), add = TRUE, border = "purple", col = NA, lwd = 3)
  plot(sf::st_geometry(x2), add = TRUE, border = "orange", col = NA, lwd = 3)
}
}
\seealso{
https://r-spatial.github.io/sf/articles/sf3.html#affine-transformations-1
}
