% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{st_xy_from_cells}
\alias{st_xy_from_cells}
\title{Given cell indices, retrieve xy coordinates for}
\usage{
st_xy_from_cells(
  x = toy_single(),
  cells = c(60L, 89L, 38L, 16L, 22L, 24L, 83L, 71L, 88L, 20L),
  where = c("start", "center", "end")[2]
)
}
\arguments{
\item{x}{stars object}

\item{cells}{numeric 1-d address (indices) of pixels of interest}

\item{where}{character, see \code{\link[stars]{st_get_dimension_values}}}
}
\value{
sf object
}
\description{
Given cell indices, retrieve xy coordinates for
}
