% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{stars_index_to_loc}
\alias{stars_index_to_loc}
\title{Convert a index (1, 2, 3,..., ncells*nbands) into stars obeject to row, col, cell and
   band location coordinates}
\usage{
stars_index_to_loc(index, x, form = c("table", "sf")[1])
}
\arguments{
\item{index}{vector of multi-band index coordinates}

\item{x}{stars object}

\item{form}{character, specifies output format as "table" (default) or "sf"}
}
\value{
a tibble of index, cell, col, row, x, y, and band
\itemize{
 \item{index, 1-based 3d index into objects if as array}
 \item{cell, 1-based 2d index into each band as if a matrix}
 \item{col, 1-based column index}
 \item{row, 1-based row index}
 \item{x, coordinate of center of cell, missing if \code{form} is 'sf'}
 \item{y, coordinate of center of cell, missing if \code{form} is 'sf'}
 \item{band, 1-based integer index of band}
}
}
\description{
Convert a index (1, 2, 3,..., ncells*nbands) into stars obeject to row, col, cell and
   band location coordinates
}
