% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{stars_pts_to_loc}
\alias{stars_pts_to_loc}
\title{Convert a sf POINTS object with a band variable into stars object to row, col,
cell x, y, and band location coordinates (cell centers).}
\usage{
stars_pts_to_loc(
  pts = dplyr::select(stars_index_to_loc(form = "sf"), dplyr::all_of("band")),
  x = toy_multi(nb = 3, nc = 5, nr = 4, mask = FALSE),
  form = c("table", "sf")[1]
)
}
\arguments{
\item{pts}{sf POINT object with one variable whose name matches the name of the band dimension
in the \code{x}.}

\item{x}{stars object.  Limited to [x,y] of [x,y, band] dimensions}

\item{form}{character, specifies output format as "table" (default) or "sf"}
}
\value{
a tibble of index, cell, col, row, x, y, and layer
\itemize{
 \item{index, 1-based 3d index into objects if as array}
 \item{cell, 1-based 2d index into each band as if a matrix}
 \item{col, 1-based column index}
 \item{row, 1-based row index}
 \item{x, coordinate of center of cell (not original input x value), missing if \code{form} is 'sf'}
 \item{y, coordinate of center of cell (not original input y value), missing if \code{form} is 'sf'}
 \item{band, 1-based integer index of band (note original band value)}
}
}
\description{
Convert a sf POINTS object with a band variable into stars object to row, col,
cell x, y, and band location coordinates (cell centers).
}
