% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy.R
\name{toy_mask}
\alias{toy_mask}
\title{Create a lightweight toy \code{stars} object}
\usage{
toy_mask(nc = 10, nr = 10)
}
\arguments{
\item{nc}{integer, number of columns}

\item{nr}{integer, number of rows}
}
\value{
stars mask with NA values assigned to area to be masked
}
\description{
Create a lightweight toy \code{stars} object
}
