% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy.R
\name{toy_multi}
\alias{toy_multi}
\title{Create a lightweight toy \code{stars} object}
\usage{
toy_multi(nc = 10, nr = 10, nb = 5)
}
\arguments{
\item{nc}{integer, number of columns}

\item{nr}{integer, number of rows}

\item{nb}{integer, number of bands}
}
\value{
stars object with regions masked
}
\description{
Create a lightweight toy \code{stars} object
}
