% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy.R
\name{toy_points}
\alias{toy_points}
\title{Create a points object suitable for use with the toy object}
\usage{
toy_points(n = 10, nc = 10, nr = 10, nb = 5)
}
\arguments{
\item{n}{numeric, the number of points}

\item{nc}{integer, number of columns in destination stars object}

\item{nr}{integer, number of rows in destination stars object}

\item{nb}{integer, number of bands in destination stars object}
}
\value{
sf POINTS object
}
\description{
Create a points object suitable for use with the toy object
}
