% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy.R
\name{toy_polygon}
\alias{toy_polygon}
\title{Create a polygon object suitable for use with the toy object}
\usage{
toy_polygon(nc = 10, nr = 10)
}
\arguments{
\item{nc}{integer, number of columns}

\item{nr}{integer, number of rows}
}
\value{
sf POLYGON object
}
\description{
Create a polygon object suitable for use with the toy object
}
