% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy.R
\name{toy_single}
\alias{toy_single}
\title{Create a single band toy \code{stars} object}
\usage{
toy_single(nc = 10, nr = 10, mask = FALSE)
}
\arguments{
\item{nc}{integer, number of columns}

\item{nr}{integer, number of rows}

\item{mask}{logical, if TRUE mask the lower left portion}
}
\value{
stars object with regions masked
}
\description{
Create a single band toy \code{stars} object
}
\seealso{
[Example from introduction](https://r-spatial.github.io/stars/articles/stars4.html#regular-grids)
}
