% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinkle.r
\name{twinkle}
\alias{twinkle}
\alias{littlestar}
\alias{estrellita}
\title{Twinkle twinkle little start}
\usage{
twinkle(n = 500, cols = c("white", "yellow", "gold"), tail = 100,
  stepsize = 0.002, deg_from = -pi/8, deg_to = pi/4, col_decay_rate = 2,
  size_decay_rate = 0.99, stargen = function(n) {     stats::rbeta(n, 2,
  5)/10 }, xlim = c(-2, 2), ylim = c(-2, 2), bg = "#041836FF")
}
\arguments{
\item{n}{Integer scalar. Number of stars to generate.}

\item{cols}{Vector of colors. Passed to \link[grDevices:colorRamp]{grDevices::colorRamp}.}

\item{tail}{Integer scalar. Length of the tails.}

\item{stepsize}{Numeric scalar.}

\item{deg_from, deg_to}{Numeric scalar. A value in radians (0 to 2\code{pi}).}

\item{col_decay_rate}{Integer scalar. The higher the value the longer the light
of the star lasts.}

\item{size_decay_rate}{Numeric scalar. Multiplied by the size of the star at each step
(change of size).}

\item{stargen}{A function. The star generator. Should receive an integer \code{n},
number of stars, and return a vector of that length with values greater than 0.}

\item{xlim, ylim}{Limits of the plotting device. This works together with \code{stargen}.}

\item{bg}{Character scalar. Passed to \link[graphics:par]{graphics::par}.}
}
\value{
NULL (invisible)
}
\description{
How I wonder what you are... Creates a nice plot with
}
\examples{

set.seed(12)
twinkle(100)
}
\author{
Mateo and George G. Vega
}
