% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinkle.R
\name{twinkle_deploy}
\alias{twinkle_deploy}
\title{ALl-in-one deploy}
\usage{
twinkle_deploy(name, production = FALSE)
}
\arguments{
\item{name}{Name of the application within the twinkle
configuration. We don't check that the application actually
exists within your configuration (or indeed even read your
configuration at all) because the application for deletion might
have been removed from the configuration already.}
}
\value{
Nothing
}
\description{
All-in-one deployment.  Runs through all steps (update source,
install packages and sync to staging and possibly to production).
This is intended either for the initial setup of an application or
a full update.  For more control use the individual functions
\code{\link[=twinkle_update_src]{twinkle_update_src()}}, \code{\link[=twinkle_install_packages]{twinkle_install_packages()}} and
\code{\link[=twinkle_sync]{twinkle_sync()}}
}
