% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinkle.R
\name{twinkle_deploy_key}
\alias{twinkle_deploy_key}
\title{Show or create a deploy key}
\usage{
twinkle_deploy_key(name, recreate = FALSE)
}
\arguments{
\item{name}{Name of the application within the twinkle configuration}

\item{recreate}{Logical, indicating if we should create the key
again even if it already exists.}
}
\value{
Invisibly, the public key as a string
}
\description{
Show (and if needed, create or recreate) a deploy key for an
private application.  Instructions will be printed to explain how
to add the key to the repository.  You can have multiple deploy
keys for a repository.
}
