% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinkle.R
\name{twinkle_list}
\alias{twinkle_list}
\title{List apps}
\usage{
twinkle_list(pattern = NULL)
}
\arguments{
\item{pattern}{Optional pattern to search for (a regular expression)}
}
\value{
A character vector
}
\description{
List known apps from the config
}
