% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinkle.R
\name{twinkle_logs}
\alias{twinkle_logs}
\title{List of view logs}
\usage{
twinkle_logs(name, list = FALSE, filename = NULL)
}
\arguments{
\item{name}{Name of the application}

\item{list}{Logical, indicating if we should list filenames of log
files, rather than show the last, or a particular, log file.}

\item{filename}{Optional filename, to show a particular log.
Otherwise we print the most recent log.}
}
\value{
A character vector
}
\description{
List or view logs from the shiny server.
}
