% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinkle.R
\name{twinkle_restart}
\alias{twinkle_restart}
\title{Restart an application}
\usage{
twinkle_restart(name, production)
}
\arguments{
\item{name}{Name of the application within the twinkle
configuration. We don't check that the application actually
exists within your configuration (or indeed even read your
configuration at all) because the application for deletion might
have been removed from the configuration already.}
}
\value{
Nothing
}
\description{
Tell shiny to restart an application.  This will trigger for
\emph{both} the staging and production versions.  This needs to be done
rarely, but would be needed in the case where the library has been
updated but the application source has not in order to force shiny
to pick up the new packages.
}
\details{
This will not work for multiple applications that are present
within a directory!  We would need to extend things so that we
knew where these different applications were (either in the
configuration file, or by applying heuristics through the source
tree) and touch a file in every sub application.  This has
historically been an issue for the old odin/shiny shortcourse app,
but that is now obsolete.  You can always restart the server
itself of course.
}
