% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinkle.R
\name{twinkle_status}
\alias{twinkle_status}
\title{Query application status}
\usage{
twinkle_status(name)
}
\arguments{
\item{name}{Name of the application}
}
\value{
Nothing
}
\description{
Query status for an application
}
