% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinkle.R
\name{twinkle_sync}
\alias{twinkle_sync}
\title{Sync app}
\usage{
twinkle_sync(name, production)
}
\arguments{
\item{name}{Name of the app}

\item{production}{Logical, indicating if we want to update the
production instance.  If \code{FALSE}, then staging is updated.}
}
\value{
A named character vector with the ids of the repo (last
SHA) and the library (conan id)
}
\description{
Sync an app and its library to the staging or production shiny
server.
}
