% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinkle.R
\name{twinkle_update_src}
\alias{twinkle_update_src}
\title{Update twinke app source}
\usage{
twinkle_update_src(name, branch = NULL)
}
\arguments{
\item{name}{Name of the app}

\item{branch}{Optionally, the branch to use (e.g., if testing a
branch on staging).  This overrides the configuration within the
application.}
}
\value{
Nothing
}
\description{
Update source for a twinkle application.
}
