% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcano_multi}
\alias{volcano_multi}
\title{Retrieve a multi-attribute stars object of Auckland’s Maunga Whau volcano topography.
Attribute/bands 2 to \code{n} are slightly altered from the original as per
\code{\link{volcano_single}}}
\usage{
volcano_multi(n = 3, what = c("attributes", "bands")[1], indexed = FALSE, ...)
}
\arguments{
\item{n}{numeric, the number attributes (or bands)}

\item{what}{character, if "attributes" (default) then yield a \code{n}-attribute object, but 
if "bands" then yield a single attribute object with \code{n} bands}

\item{indexed}{logical, if TRUE then assign 1,2,3,... as cell values}

\item{...}{other arguments for \code{\link{volcano_single}}, especially \code{threshold}.
Ignored if \code{indexed} is \code{TRUE}.}
}
\value{
stars class object
}
\description{
Retrieve a multi-attribute stars object of Auckland’s Maunga Whau volcano topography.
Attribute/bands 2 to \code{n} are slightly altered from the original as per
\code{\link{volcano_single}}
}
\seealso{
\code{\link{volcano_single}}
}
