% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcano_points}
\alias{volcano_points}
\title{Generate a table of random points in a \code{stars} object}
\usage{
volcano_points(x = volcano_multi(what = "bands"), ...)
}
\arguments{
\item{x}{stars object}

\item{...}{further arguments for \link{random_points}}
}
\value{
tibble of locations with values
}
\description{
Generate a table of random points in a \code{stars} object
}
