% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcano_single}
\alias{volcano_single}
\title{Retrieve a stars object of Auckland’s Maunga Whau volcano topography.  Adapted from
USGS-R inlmisc package.}
\usage{
volcano_single(indexed = FALSE, name = "values")
}
\arguments{
\item{indexed}{logical, if TRUE then assign 1,2,3,... as cell values}

\item{name}{charcater, the name to apply to the attribute}
}
\value{
single attribute stars
}
\description{
Retrieve a stars object of Auckland’s Maunga Whau volcano topography.  Adapted from
USGS-R inlmisc package.
}
\seealso{
\url{https://waterdata.usgs.gov/blog/inlmiscmaps/}

\url{https://CRAN.R-project.org/package=inlmisc}
}
