% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{write_gpkg}
\alias{write_gpkg}
\title{Write to a geopackage file by writing to a temp file first and then
copying to the specified destination.}
\usage{
write_gpkg(obj, dsn, overwrite = FALSE, ...)
}
\arguments{
\item{obj}{object to write}

\item{dsn}{character, the file to write to}

\item{overwrite, }{logical, overwite existing file?}

\item{...}{other arguments passed to \code{\link[sf]{write_sf}}}
}
\value{
obj invisibly
}
\description{
Write to a geopackage file by writing to a temp file first and then
copying to the specified destination.
}
\seealso{
\href{https://github.com/r-spatial/sf/issues/628#issuecomment-381690663}{sf issue 628}
}
