% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs.R
\name{pairs,data.frame-method}
\alias{pairs,data.frame-method}
\title{Pairwise matrix of scatterplot for data.frame objects - modeled after `pairs`
function in the [terra](https://CRAN.R-project.org/package=terra) package.}
\usage{
\S4method{pairs}{data.frame}(x, hist = TRUE, cor = TRUE, use = "pairwise.complete.obs", ...)
}
\arguments{
\item{x}{SpatRaster}

\item{hist}{logical. If TRUE a histogram of the values is shown on the diagonal}

\item{cor}{logical. If TRUE the correlation coefficient is shown in the upper panels}

\item{use}{argument passed to the \code{\link[stats]{cor}} function}

\item{...}{additional arguments (graphical parameters)}
}
\value{
a pairs plot
}
\description{
I first made a `pairs-stars` for the [tidysdm](https://CRAN.R-project.org/package=tidysdm)to match that 
offered by [terra](https://CRAN.R-project.org/package=terra).  But we like the plot so much we thought it worth
adding a data.frame method.
}
\examples{
x = iris |>
  dplyr::as_tibble() |>
  dplyr::select(-dplyr::all_of("Species"))
pairs(x)
}
