% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinkle-package.R
\docType{package}
\name{twinkle}
\alias{twinkle}
\title{twinkle: Tools for working with [https://CRAN.R-project.org/package=stars] and [simple features](https://CRAN.R-project.org/package=sf).}
\description{
Tools for working with [https://CRAN.R-project.org/package=stars] and [simple features](https://CRAN.R-project.org/package=sf).
}
