% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinkle.R
\name{twinkle_delete_app}
\alias{twinkle_delete_app}
\title{Delete an application}
\usage{
twinkle_delete_app(name, production = FALSE)
}
\arguments{
\item{name}{Name of the application within the twinkle
configuration. We don't check that the application actually
exists within your configuration (or indeed even read your
configuration at all) because the application for deletion might
have been removed from the configuration already.}

\item{production}{Logical, indicating if we should \emph{additionally}
delete on production too. Otherwise production is left alone.
This allows you to totally redeploy an application but leave it
running until you are happy with the version, then sync it into
place.}
}
\value{
Nothing
}
\description{
Delete an application.  This removes everything associated with
the application (depending on how far through deployment you might
have been); the real and staging instances, the library, the
deploy key (if private) and the source code.  You might want to
use this when deleting an application that is no longer needed, or
if something untoward has happened and it would be convenient to
start from a clean state.
}
