% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinkle.R
\name{twinkle_history}
\alias{twinkle_history}
\title{Show application history}
\usage{
twinkle_history(name)
}
\arguments{
\item{name}{Name of the application}
}
\value{
Nothing
}
\description{
Show history of an application
}
