% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinkle.R
\name{twinkle_install_packages}
\alias{twinkle_install_packages}
\title{Install app packages}
\usage{
twinkle_install_packages(name)
}
\arguments{
\item{name}{Name of the app}
}
\value{
Nothing
}
\description{
Install packages for an app
}
